+++ public/tier1/utlhashtable.h
@@ -368,7 +368,7 @@ void CUtlHashtable<KeyT, ValueT, KeyHashT, KeyIsEqualT, AltKeyT>::BumpEntry( uns
 
 	entry_t* table = m_table.Base();
 	unsigned int slotmask = m_table.Count()-1;
-	unsigned int new_flags_and_hash = table[idx].flags_and_hash & (FLAG_LAST | MASK_HASH);
+	unsigned int new_flags_and_hash = table[idx].flags_and_hash & ((unsigned int)FLAG_LAST | (unsigned int)MASK_HASH);
 
 	unsigned int chainid = entry_t::IdealIndex( new_flags_and_hash, slotmask );
